React/backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
numpy
pandas
matplotlib
scikit-learn
